requirements.txt
livekit-agents[aws,openai,cartesia,silero,turn-detector]~=1.0
livekit-plugins-noise-cancellation~=0.2
python-dotenv